Clean up instructor and lesson screens
