+++ src/screens/Instructors/screens/Instructor/components/GetPublishedWidget/components/Checklist/components/MoreInfo.js
