+++ src/screens/Instructors/screens/Instructor/components/GetPublishedWidget/utils/isStepComplete/index.js
