+++ src/screens/Instructors/screens/Instructor/components/HelpWidget/index.js
@@ -33,16 +33,8 @@ const items=[
   },
 ]
 
-export default ({instructor}) => {
-
-  const {published_lessons} = instructor
-
-  if(published_lessons === 0) {
-    return null
-  }
-
-  return (
-    <Widget title={helpTitleText}>
+export default ({publishedLessons}) => publishedLessons === 0
+  ? <Widget title={helpTitleText}>
       <List items={map(items, (item, index) => (
         <div key={index}>
           <Paragraph>
@@ -56,5 +48,4 @@ export default ({instructor}) => {
         </div>
       ))} />
     </Widget>
-  )
-}
+  : null
