+++ src/screens/Instructors/screens/Instructor/components/RevenueWidget/utils/removeRevenueMonth/index.test.js
